fix(dto): align DTOs and OpenAPI specs after validation review (#212)

+++ src/dto/dashboard.rs
@@ -9,8 +9,8 @@ use serde::Serialize;
 pub struct CurrentPeriodResponse {
     pub spent: i64,
     pub target: i64,
-    pub days_remaining: i32,
-    pub days_in_period: i32,
+    pub days_remaining: i64,
+    pub days_in_period: i64,
     pub projected_spend: i64,
 }
 
@@ -21,7 +21,7 @@ pub struct CurrentPeriodResponse {
 pub struct NetPositionResponse {
     pub total: i64,
     pub difference_this_period: i64,
-    pub number_of_accounts: i32,
+    pub number_of_accounts: i64,
     pub liquid_amount: i64,
     pub protected_amount: i64,
     pub debt_amount: i64,
@@ -32,7 +32,7 @@ pub struct NetPositionResponse {
 #[derive(Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct BudgetStabilityResponse {
-    pub stability: i32,
-    pub periods_within_range: i32,
+    pub stability: i64,
+    pub periods_within_range: i64,
     pub periods_stability: Vec<bool>,
 }
